requirements.txt
Django==1.6.1
pinax-theme-bootstrap==5.0.1
django-user-accounts==1.0c1
metron==1.3.1
eventlog==0.7.0
django-jsonfield==0.9.12
django-forms-bootstrap
south
django-tastypie
django-bootstrap3-datetimepicker
